Refactor code editor page into smaller components
