+++ frontend/src/components/QuestionContainer/QuestionContainer.scss
@@ -1,0 +1,21 @@
+.question-container {
+    margin-bottom: 20px;
+  
+    .question-title {
+      margin-bottom: 10px;
+      color: #e4e4e4;
+      display: flex;
+      align-items: center;
+    }
+  
+    .details {
+      display: flex;
+      gap: 10px;
+  
+      .light-grey-chip {
+        background-color: #373737;
+        color: #ffffff;
+        border-radius: 6px;
+      }
+    }
+  }
